functions/iam_monitor/requirements.txt
functions-framework>=3.0.0
google-cloud-pubsub>=2.18.0
google-cloud-logging>=3.8.0
google-cloud-secret-manager>=2.16.0
google-cloud-storage>=2.10.0
boto3>=1.28.0
botocore>=1.31.0
pyyaml>=6.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
slack-sdk>=3.21.0
requests>=2.31.0
tenacity>=8.2.0
python-dotenv>=1.0.0
